Fix test and implementation
