+++ rs/nervous_system/integration_tests/BUILD.bazel
@@ -35,6 +35,7 @@ BASE_DEPENDENCIES = [
     "@crate_index//:assert_matches",
     "@crate_index//:candid",
     "@crate_index//:futures",
+    "@crate_index//:hex",
     "@crate_index//:ic-management-canister-types",
     "@crate_index//:itertools",
     "@crate_index//:lazy_static",
